requirements.txt
crewai
crewai[tools]
python-dotenv
composio_core
composio-crewai
langchain-google-genai
discord
tinydb
langchain_core==0.2.3